requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-MySQLdb==2.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
packaging==24.0
PyJWT==2.8.0
PyMonCtl==0.7
pywin32==306
PyWinBox==0.6
PyWinCtl==0.3
typing_extensions==4.11.0
Werkzeug==3.0.2